Remove HeaderBuilder
MapperCompiler was an anemic object with zero behavior and HeaderBuilder
was pretty much reduntant so all its logic was moved to MapperCompiler
which simplifies the design
