+++ repository/spec/integration/command_macros_spec.rb
@@ -99,7 +99,7 @@ RSpec.describe ROM::Repository, '.command' do
     struct_definition = [:users, [[:attribute, repo.users.schema[:id]],
                                   [:attribute, repo.users.schema[:name]]]]
 
-    expect(user).to be_an_instance_of repo.users.mappers.compiler.header_builder.struct_builder.cache[struct_definition.hash]
+    expect(user).to be_an_instance_of repo.users.mappers.compiler.struct_builder.cache[struct_definition.hash]
   end
 
   describe 'using plugins' do
